Use generic WordPress compliance gates
