+++ .github/workflows/plugin-check.yml
@@ -1,0 +1,58 @@
+name: WordPress Plugin Check
+
+on:
+    pull_request:
+    push:
+        branches:
+            - master
+            - develop
+            - 'release/**'
+
+permissions:
+    contents: read
+
+concurrency:
+    group: plugin-check-${{ github.ref }}
+    cancel-in-progress: true
+
+jobs:
+    plugin-check:
+        name: Check assembled plugin
+        runs-on: ubuntu-latest
+
+        steps:
+            - name: Checkout
+              uses: actions/checkout@v4
+
+            - name: Setup PHP
+              uses: shivammathur/setup-php@v2
+              with:
+                  php-version: '8.3'
+                  tools: composer:v2
+                  coverage: none
+
+            - name: Setup Node.js
+              uses: actions/setup-node@v4
+              with:
+                  node-version: '16'
+                  cache: npm
+
+            - name: Install dependencies
+              run: npm ci
+
+            - name: Assemble release plugin
+              env:
+                  COMPOSER_AUTH: '{}'
+                  COMPOSER_HOME: ${{ runner.temp }}/composer-release
+              run: npm run release:build
+
+            - name: Run WordPress Plugin Check
+              uses: wordpress/plugin-check-action@v1
+              with:
+                  build-dir: ./build
+                  slug: content-control
+                  exclude-checks: plugin_review_phpcs
+                  exclude-directories: |
+                      vendor
+                      vendor-prefixed
+                  ignore-warnings: true
