--- .github/workflows/wporg-source-compliance.yml
@@ -1,28 +1,0 @@
-name: WordPress.org Source Compliance
-
-on:
-    pull_request:
-    push:
-        branches:
-            - master
-            - develop
-
-permissions:
-    contents: read
-
-jobs:
-    verify-source:
-        name: Verify product source
-        runs-on: ubuntu-latest
-
-        steps:
-            - name: Checkout
-              uses: actions/checkout@v4
-
-            - name: Setup Node.js
-              uses: actions/setup-node@v4
-              with:
-                  node-version: '20'
-
-            - name: Reject private plugin-delivery implementation
-              run: node bin/verify-wporg-artifact.js --source .
