+++ package.json
@@ -95,10 +95,10 @@
 		"pretest:unit:php": "wp-env start",
 		"test:unit:php": "wp-env run phpunit /var/www/html/wp-content/plugins/content-control/vendor/bin/phpunit -c /var/www/html/wp-content/plugins/content-control/phpunit.xml.dist --verbose",
 		"translate": "wp-pot --src './classes/**/*.php' --src '!./build/**/*.*' --src '!./vendor/**/*.*' --src '!./dist/**/*.*' --write=true --dest-file 'languages/content-control.pot' --package 'Content Control' --domain 'content-control' --last-translator 'Code Atlantic <support@code-atlantic.com>' --team 'Code Atlantic <support@code-atlantic.com>' --bug-report 'https://code-atlantic.com/contact/'",
-		"verify:wporg:source": "node bin/verify-wporg-artifact.js --source .",
-		"verify:wporg:build": "node bin/verify-wporg-artifact.js --release-tree build",
-		"release:build": "composer install -o --no-dev --classmap-authoritative && npm run build:release && npm run verify:wporg:source && npm run copy-build && npm run verify:wporg:build",
-		"release:zip": "mv build content-control && zip -r content-control_$(node -p \"require('./package.json').version\").zip content-control && rm -rf content-control && node bin/verify-wporg-artifact.js content-control_$(node -p \"require('./package.json').version\").zip",
+		"verify:release-artifact": "node bin/verify-release-artifact.js",
+		"verify:release-tree": "node bin/verify-release-artifact.js --release-tree build",
+		"release:build": "composer install -o --no-dev --classmap-authoritative && npm run build:release && npm run copy-build && npm run verify:release-tree",
+		"release:zip": "mv build content-control && zip -r content-control_$(node -p \"require('./package.json').version\").zip content-control && rm -rf content-control && node bin/verify-release-artifact.js content-control_$(node -p \"require('./package.json').version\").zip",
 		"release:start": "git flow release start",
 		"release:finish": "git flow release finish",
 		"release:publish": "git push origin master --tags && git push origin master",
