+++ tests/unit/bin/verify-release-artifact.test.js
@@ -6,15 +6,15 @@ const { execFileSync } = require( 'child_process' );
 const {
 	verifyArchiveEntryTypes,
 	verifyArchiveStructure,
-	verifyArtifactEntries,
 	verifyReleaseManifest,
+	verifyReleaseTree,
 	verifyZip,
-} = require( '../../../bin/verify-wporg-artifact' );
+} = require( '../../../bin/verify-release-artifact' );
 
-describe( 'WordPress.org artifact verifier', () => {
-	test( 'accepts the runtime Composer files in a release ZIP', () => {
+describe( 'release artifact verifier', () => {
+	test( 'accepts the assembled runtime files in a release ZIP', () => {
 		const temporaryRoot = fs.mkdtempSync(
-			path.join( os.tmpdir(), 'content-control-wporg-verifier-' )
+			path.join( os.tmpdir(), 'content-control-release-verifier-' )
 		);
 		const pluginRoot = path.join( temporaryRoot, 'content-control' );
 		const zipPath = path.join( temporaryRoot, 'content-control.zip' );
@@ -22,8 +22,9 @@ describe( 'WordPress.org artifact verifier', () => {
 		try {
 			for ( const relativePath of [
 				'content-control.php',
-				'classes/Plugin/License.php',
-				'classes/RestAPI/License.php',
+				'readme.txt',
+				'dist/settings-page.js',
+				'dist/settings-page.css',
 				'vendor/autoload.php',
 				'vendor/composer/autoload_real.php',
 				'vendor/composer/installed.json',
@@ -47,15 +48,15 @@ describe( 'WordPress.org artifact verifier', () => {
 		expect(
 			verifyArchiveStructure( [
 				'content-control/',
-				'content-control/classes/Plugin/License.php',
-				'content-control/classes/Plugin/License.php',
-				'decoy/classes/Plugin/License.php',
+				'content-control/classes/Rules.php',
+				'content-control/classes/Rules.php',
+				'decoy/classes/Rules.php',
 				'content-control/../private.php',
 			] )
 		).toEqual(
 			expect.arrayContaining( [
-				'duplicate archive path: content-control/classes/Plugin/License.php',
-				'unexpected archive root: decoy/classes/Plugin/License.php',
+				'duplicate archive path: content-control/classes/Rules.php',
+				'unexpected archive root: decoy/classes/Rules.php',
 				'unsafe archive path: content-control/../private.php',
 			] )
 		);
@@ -80,9 +81,30 @@ describe( 'WordPress.org artifact verifier', () => {
 		);
 	} );
 
-	test( 'allows runtime Composer records but rejects other vendor files', () => {
+	test( 'requires runtime files and rejects repository-only paths', () => {
+		const failures = verifyReleaseManifest( [
+			'content-control/content-control.php',
+			'content-control/readme.txt',
+			'content-control/vendor/autoload.php',
+			'content-control/tests/fixture.php',
+		] );
+
+		expect( failures ).toEqual(
+			expect.arrayContaining( [
+				'missing release path: dist/settings-page.js',
+				'missing release path: dist/settings-page.css',
+				'forbidden release path: tests/fixture.php',
+			] )
+		);
+	} );
+
+	test( 'allows only the runtime Composer records copied by the build', () => {
 		expect(
 			verifyReleaseManifest( [
+				'content-control/content-control.php',
+				'content-control/readme.txt',
+				'content-control/dist/settings-page.js',
+				'content-control/dist/settings-page.css',
 				'content-control/vendor/',
 				'content-control/vendor/autoload.php',
 				'content-control/vendor/composer/',
@@ -94,43 +116,31 @@ describe( 'WordPress.org artifact verifier', () => {
 		expect(
 			verifyReleaseManifest( [
 				'content-control/vendor/library/source.php',
-				'content-control/tests/fixture.php',
 			] )
-		).toEqual(
-			expect.arrayContaining( [
-				'forbidden release path: vendor/library/source.php',
-				'forbidden release path: tests/fixture.php',
-			] )
+		).toContain( 'forbidden release path: vendor/library/source.php' );
+	} );
+
+	test( 'verifies an assembled release tree without inspecting implementation names', () => {
+		const temporaryRoot = fs.mkdtempSync(
+			path.join( os.tmpdir(), 'content-control-release-tree-' )
 		);
-	} );
 
-	test( 'allows the narrow old-Pro license field in public Core assets', () => {
-		const contents = {
-			'classes/Plugin/License.php': '<?php // Compatibility service.',
-			'classes/RestAPI/License.php': '<?php // Compatibility controller.',
-			'dist/settings-page.js':
-				'Paste or enter your license key; activateLicense(); deactivateLicense();',
-		};
+		try {
+			for ( const relativePath of [
+				'content-control.php',
+				'readme.txt',
+				'dist/settings-page.js',
+				'dist/settings-page.css',
+				'vendor/autoload.php',
+			] ) {
+				const filePath = path.join( temporaryRoot, relativePath );
+				fs.mkdirSync( path.dirname( filePath ), { recursive: true } );
+				fs.writeFileSync( filePath, 'release fixture' );
+			}
 
-		expect(
-			verifyArtifactEntries( Object.keys( contents ), ( entry ) =>
-				Buffer.from( contents[ entry ] )
-			)
-		).toEqual( [] );
-	} );
-
-	test( 'still rejects installer behavior beside the compatibility field', () => {
-		const contents = {
-			'classes/Plugin/License.php': '<?php // Compatibility service.',
-			'classes/RestAPI/License.php': '<?php // Compatibility controller.',
-			'dist/settings-page.js':
-				'Paste or enter your license key; activatePro();',
-		};
-
-		expect(
-			verifyArtifactEntries( Object.keys( contents ), ( entry ) =>
-				Buffer.from( contents[ entry ] )
-			)
-		).toContain( 'installer-specific UI in dist/settings-page.js' );
+			expect( verifyReleaseTree( temporaryRoot ).failures ).toEqual( [] );
+		} finally {
+			fs.rmSync( temporaryRoot, { recursive: true, force: true } );
+		}
 	} );
 } );
